Move image url util to components folder
